requirements.txt
google-cloud-aiplatform
sentence-transformers
tiktoken 
chromadb
langchain
langchain-community
langchain-google-vertexai
flask
GitPython
python-dotenv